修改更新

+++ utilcode/src/main/java/com/blankj/utilcode/util/CrashUtils.java
@@ -1,29 +1,37 @@
 package com.blankj.utilcode.util;
 
+import android.annotation.SuppressLint;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.os.Build;
 import android.os.Environment;
 import android.support.annotation.NonNull;
+import android.support.annotation.RequiresPermission;
+import android.util.Log;
 
+import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.io.PrintWriter;
+import java.io.StringWriter;
 import java.lang.Thread.UncaughtExceptionHandler;
 import java.text.Format;
 import java.text.SimpleDateFormat;
 import java.util.Date;
-import java.util.Locale;
-import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Callable;
+import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Executors;
+import java.util.concurrent.Future;
+
+import static android.Manifest.permission.WRITE_EXTERNAL_STORAGE;
 
 /**
  * <pre>
  *     author: Blankj
  *     blog  : http://blankj.com
  *     time  : 2016/09/27
- *     desc  : 崩溃相关工具类
+ *     desc  : utils about crash
  * </pre>
  */
 public final class CrashUtils {
@@ -33,19 +41,20 @@ public final class CrashUtils {
     private static String versionName;
     private static int    versionCode;
 
-    private static ExecutorService sExecutor;
-
     private static final String FILE_SEP = System.getProperty("file.separator");
-    private static final Format FORMAT   = new SimpleDateFormat("MM-dd HH-mm-ss", Locale.getDefault());
-
-    private static final String CRASH_HEAD;
+    @SuppressLint("SimpleDateFormat")
+    private static final Format FORMAT   = new SimpleDateFormat("MM-dd HH-mm-ss");
 
     private static final UncaughtExceptionHandler DEFAULT_UNCAUGHT_EXCEPTION_HANDLER;
     private static final UncaughtExceptionHandler UNCAUGHT_EXCEPTION_HANDLER;
 
+    private static OnCrashListener sOnCrashListener;
+
     static {
         try {
-            PackageInfo pi = Utils.getApp().getPackageManager().getPackageInfo(Utils.getApp().getPackageName(), 0);
+            PackageInfo pi = Utils.getApp()
+                    .getPackageManager()
+                    .getPackageInfo(Utils.getApp().getPackageName(), 0);
             if (pi != null) {
                 versionName = pi.versionName;
                 versionCode = pi.versionCode;
@@ -53,55 +62,55 @@ public final class CrashUtils {
         } catch (PackageManager.NameNotFoundException e) {
             e.printStackTrace();
         }
-
-        CRASH_HEAD = "\n************* Crash Log Head ****************" +
-                "\nDevice Manufacturer: " + Build.MANUFACTURER +// 设备厂商
-                "\nDevice Model       : " + Build.MODEL +// 设备型号
-                "\nAndroid Version    : " + Build.VERSION.RELEASE +// 系统版本
-                "\nAndroid SDK        : " + Build.VERSION.SDK_INT +// SDK版本
-                "\nApp VersionName    : " + versionName +
-                "\nApp VersionCode    : " + versionCode +
-                "\n************* Crash Log Head ****************\n\n";
-
         DEFAULT_UNCAUGHT_EXCEPTION_HANDLER = Thread.getDefaultUncaughtExceptionHandler();
 
         UNCAUGHT_EXCEPTION_HANDLER = new UncaughtExceptionHandler() {
             @Override
             public void uncaughtException(final Thread t, final Throwable e) {
                 if (e == null) {
-                    android.os.Process.killProcess(android.os.Process.myPid());
-                    System.exit(0);
+                    if (DEFAULT_UNCAUGHT_EXCEPTION_HANDLER != null) {
+                        DEFAULT_UNCAUGHT_EXCEPTION_HANDLER.uncaughtException(t, null);
+                    } else {
+                        android.os.Process.killProcess(android.os.Process.myPid());
+                        System.exit(1);
+                    }
                     return;
                 }
-                Date now = new Date(System.currentTimeMillis());
-                String fileName = FORMAT.format(now) + ".txt";
-                final String fullPath = (dir == null ? defaultDir : dir) + fileName;
-                if (!createOrExistsFile(fullPath)) return;
-                if (sExecutor == null) {
-                    sExecutor = Executors.newSingleThreadExecutor();
+
+                final String time = FORMAT.format(new Date(System.currentTimeMillis()));
+                final StringBuilder sb = new StringBuilder();
+                final String head = "************* Log Head ****************" +
+                        "\nTime Of Crash      : " + time +
+                        "\nDevice Manufacturer: " + Build.MANUFACTURER +
+                        "\nDevice Model       : " + Build.MODEL +
+                        "\nAndroid Version    : " + Build.VERSION.RELEASE +
+                        "\nAndroid SDK        : " + Build.VERSION.SDK_INT +
+                        "\nApp VersionName    : " + versionName +
+                        "\nApp VersionCode    : " + versionCode +
+                        "\n************* Log Head ****************\n\n";
+                sb.append(head);
+                StringWriter sw = new StringWriter();
+                PrintWriter pw = new PrintWriter(sw);
+                e.printStackTrace(pw);
+                Throwable cause = e.getCause();
+                while (cause != null) {
+                    cause.printStackTrace(pw);
+                    cause = cause.getCause();
                 }
-                sExecutor.execute(new Runnable() {
-                    @Override
-                    public void run() {
-                        PrintWriter pw = null;
-                        try {
-                            pw = new PrintWriter(new FileWriter(fullPath, false));
-                            pw.write(CRASH_HEAD);
-                            e.printStackTrace(pw);
-                            Throwable cause = e.getCause();
-                            while (cause != null) {
-                                cause.printStackTrace(pw);
-                                cause = cause.getCause();
-                            }
-                        } catch (IOException e) {
-                            e.printStackTrace();
-                        } finally {
-                            if (pw != null) {
-                                pw.close();
-                            }
-                        }
-                    }
-                });
+                pw.flush();
+                sb.append(sw.toString());
+                final String crashInfo = sb.toString();
+                final String fullPath = (dir == null ? defaultDir : dir) + time + ".txt";
+                if (createOrExistsFile(fullPath)) {
+                    input2File(crashInfo, fullPath);
+                } else {
+                    Log.e("CrashUtils", "create " + fullPath + " failed!");
+                }
+
+                if (sOnCrashListener != null) {
+                    sOnCrashListener.onCrash(crashInfo, e);
+                }
+
                 if (DEFAULT_UNCAUGHT_EXCEPTION_HANDLER != null) {
                     DEFAULT_UNCAUGHT_EXCEPTION_HANDLER.uncaughtException(t, e);
                 }
@@ -114,34 +123,78 @@ public final class CrashUtils {
     }
 
     /**
-     * 初始化
-     * <p>需添加权限 {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>}</p>
+     * Initialization.
+     * <p>Must hold
+     * {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />}</p>
      */
+    @RequiresPermission(WRITE_EXTERNAL_STORAGE)
     public static void init() {
         init("");
     }
 
     /**
-     * 初始化
-     * <p>需添加权限 {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>}</p>
+     * Initialization
+     * <p>Must hold
+     * {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />}</p>
      *
-     * @param crashDir 崩溃文件存储目录
+     * @param crashDir The directory of saving crash information.
      */
+    @RequiresPermission(WRITE_EXTERNAL_STORAGE)
     public static void init(@NonNull final File crashDir) {
-        init(crashDir.getAbsolutePath());
+        init(crashDir.getAbsolutePath(), null);
     }
 
     /**
-     * 初始化
-     * <p>需添加权限 {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>}</p>
+     * Initialization
+     * <p>Must hold
+     * {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />}</p>
      *
-     * @param crashDir 崩溃文件存储目录
+     * @param crashDirPath The directory's path of saving crash information.
      */
-    public static void init(final String crashDir) {
-        if (isSpace(crashDir)) {
+    @RequiresPermission(WRITE_EXTERNAL_STORAGE)
+    public static void init(final String crashDirPath) {
+        init(crashDirPath, null);
+    }
+
+    /**
+     * Initialization
+     * <p>Must hold
+     * {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />}</p>
+     *
+     * @param onCrashListener The crash listener.
+     */
+    @RequiresPermission(WRITE_EXTERNAL_STORAGE)
+    public static void init(final OnCrashListener onCrashListener) {
+        init("", onCrashListener);
+    }
+
+    /**
+     * Initialization
+     * <p>Must hold
+     * {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />}</p>
+     *
+     * @param crashDir        The directory of saving crash information.
+     * @param onCrashListener The crash listener.
+     */
+    @RequiresPermission(WRITE_EXTERNAL_STORAGE)
+    public static void init(@NonNull final File crashDir, final OnCrashListener onCrashListener) {
+        init(crashDir.getAbsolutePath(), onCrashListener);
+    }
+
+    /**
+     * Initialization
+     * <p>Must hold
+     * {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />}</p>
+     *
+     * @param crashDirPath    The directory's path of saving crash information.
+     * @param onCrashListener The crash listener.
+     */
+    @RequiresPermission(WRITE_EXTERNAL_STORAGE)
+    public static void init(final String crashDirPath, final OnCrashListener onCrashListener) {
+        if (isSpace(crashDirPath)) {
             dir = null;
         } else {
-            dir = crashDir.endsWith(FILE_SEP) ? crashDir : crashDir + FILE_SEP;
+            dir = crashDirPath.endsWith(FILE_SEP) ? crashDirPath : crashDirPath + FILE_SEP;
         }
         if (Environment.MEDIA_MOUNTED.equals(Environment.getExternalStorageState())
                 && Utils.getApp().getExternalCacheDir() != null)
@@ -149,7 +202,53 @@ public final class CrashUtils {
         else {
             defaultDir = Utils.getApp().getCacheDir() + FILE_SEP + "crash" + FILE_SEP;
         }
+        sOnCrashListener = onCrashListener;
         Thread.setDefaultUncaughtExceptionHandler(UNCAUGHT_EXCEPTION_HANDLER);
+    }
+
+    ///////////////////////////////////////////////////////////////////////////
+    // interface
+    ///////////////////////////////////////////////////////////////////////////
+
+    public interface OnCrashListener {
+        void onCrash(String crashInfo, Throwable e);
+    }
+
+    ///////////////////////////////////////////////////////////////////////////
+    // other utils methods
+    ///////////////////////////////////////////////////////////////////////////
+
+    private static void input2File(final String input, final String filePath) {
+        Future<Boolean> submit = Executors.newSingleThreadExecutor().submit(new Callable<Boolean>() {
+            @Override
+            public Boolean call() {
+                BufferedWriter bw = null;
+                try {
+                    bw = new BufferedWriter(new FileWriter(filePath, true));
+                    bw.write(input);
+                    return true;
+                } catch (IOException e) {
+                    e.printStackTrace();
+                    return false;
+                } finally {
+                    try {
+                        if (bw != null) {
+                            bw.close();
+                        }
+                    } catch (IOException e) {
+                        e.printStackTrace();
+                    }
+                }
+            }
+        });
+        try {
+            if (submit.get()) return;
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        } catch (ExecutionException e) {
+            e.printStackTrace();
+        }
+        Log.e("CrashUtils", "write crash info to " + filePath + " failed!");
     }
 
     private static boolean createOrExistsFile(final String filePath) {
